pyproject.toml
[project]
name = "evening-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.4",
    "trafilatura>=2.0.0",
    "tqdm>=4.66.4",
    "ebooklib>=0.19",
    "Pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
